# specs/005-blog-content-management/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Blog Content Management API Delta
  version: 1.0.0
  description: >-
    Private blog-module additions for feature 005. Existing public post and RSS
    contracts remain unchanged and are intentionally not expanded here.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
  /posts:
    get:
      summary: List owned posts with blog filters
      parameters:
        - { $ref: '#/components/parameters/SearchQuery' }
        - { $ref: '#/components/parameters/ContentClass' }
        - { $ref: '#/components/parameters/ContentTypeId' }
        - { $ref: '#/components/parameters/ContentStatus' }
        - { $ref: '#/components/parameters/AiDisplayStatus' }
        - { $ref: '#/components/parameters/SourceType' }
        - { $ref: '#/components/parameters/CategoryId' }
        - { $ref: '#/components/parameters/TagId' }
        - { $ref: '#/components/parameters/KeywordId' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200':
          description: Cursor page of owned posts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostPage' }
    post:
      summary: Create an empty owned draft and first revision
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PostCreate' }
      responses:
        '201':
          description: Durably saved draft
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Post' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /posts/{post_id}:
    get:
      summary: Get one owned current article and its organization summary
      parameters: [{ $ref: '#/components/parameters/PostId' }]
      responses:
        '200':
          description: Current article
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Post' }
        '404': { $ref: '#/components/responses/Problem' }
    patch:
      summary: Save user edits with optimistic concurrency and create a revision
      parameters: [{ $ref: '#/components/parameters/PostId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PostPatch' }
      responses:
        '200':
          description: Saved current article
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Post' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
    delete:
      summary: Soft-delete an owned non-published article
      parameters: [{ $ref: '#/components/parameters/PostId' }]
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
...
  /blog/taxonomy/{kind}:
    get:
      summary: List categories, tags or keywords with usage counts
      parameters:
        - { $ref: '#/components/parameters/TaxonomyKind' }
        - in: query
          name: enabled
          schema: { type: boolean }
      responses:
        '200':
          description: Taxonomy items
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TaxonomyItem' }
    post:
      summary: Create a category, tag or keyword
      parameters: [{ $ref: '#/components/parameters/TaxonomyKind' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaxonomyWrite' }
      responses:
        '201':
          description: Created item
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaxonomyItem' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: access_token
  parameters:
    PostId: { in: path, name: post_id, required: true, schema: { type: string, format: uuid } }
...
    TaxonomyKind: { in: path, name: kind, required: true, schema: { type: string, enum: [category, tag, keyword] } }
    SearchQuery: { in: query, name: q, schema: { type: string, maxLength: 500 } }
...
    ContentClass: { in: query, name: content_class, schema: { type: string, maxLength: 32 } }
    ContentTypeId: { in: query, name: content_type_id, schema: { type: string, format: uuid } }
    ContentStatus: { in: query, name: content_status, schema: { type: string, maxLength: 32 } }
    AiDisplayStatus: { in: query, name: ai_status, schema: { type: string, maxLength: 32 } }
    SourceType: { in: query, name: source_type, schema: { type: string, maxLength: 24 } }
    CategoryId: { in: query, name: category_id, schema: { type: string, format: uuid } }
    TagId: { in: query, name: tag_id, schema: { type: string, format: uuid } }
    KeywordId: { in: query, name: keyword_id, schema: { type: string, format: uuid } }
    Cursor: { in: query, name: cursor, schema: { type: string, maxLength: 500 } }
  responses:
    Problem:
      description: RFC 9457 problem details; never contains private raw content
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Post:
      type: object
      additionalProperties: false
      required: [id, title, markdown, status, content_status, content_class, language, structured_data, version, current_revision_id, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        title: { type: string, maxLength: 240 }
        subtitle: { type: [string, 'null'], maxLength: 240 }
        summary: { type: [string, 'null'], maxLength: 2000 }
        markdown: { type: string, maxLength: 200000 }
        status: { type: string, enum: [draft, private, published] }
        content_status: { $ref: '#/components/schemas/ContentStatus' }
        content_class: { type: string, maxLength: 32 }
        content_type_id: { type: [string, 'null'], format: uuid }
        category_id: { type: [string, 'null'], format: uuid }
        tag_ids: { type: array, items: { type: string, format: uuid }, maxItems: 50 }
        keyword_ids: { type: array, items: { type: string, format: uuid }, maxItems: 100 }
        language: { type: string, maxLength: 16 }
        editor_mode: { type: string, enum: [markdown, rich, split] }
        occurred_at: { type: [string, 'null'], format: date-time }
        location: { type: [string, 'null'], maxLength: 240 }
        project: { type: [string, 'null'], maxLength: 240 }
        structured_data: { type: object, maxProperties: 100 }
        source_summary: { type: array, items: { $ref: '#/components/schemas/PostSourceSummary' } }
        ai_summary: { $ref: '#/components/schemas/PostAiSummary' }
        version: { type: integer, minimum: 1 }
        current_revision_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PostCreate:
      type: object
      additionalProperties: false
      properties:
        title: { type: string, default: 未命名文章, maxLength: 240 }
        markdown: { type: string, default: '', maxLength: 200000 }
        content_class: { type: string, default: essay }
        content_type_id: { type: [string, 'null'], format: uuid }
        ai_enabled: { type: boolean, default: false }
        skill_id: { type: [string, 'null'], format: uuid }
        save_as_defaults: { type: boolean, default: false }
    PostPatch:
      type: object
      additionalProperties: false
      required: [version]
      properties:
        version: { type: integer, minimum: 1 }
        title: { type: string, maxLength: 240 }
        subtitle: { type: [string, 'null'], maxLength: 240 }
        summary: { type: [string, 'null'], maxLength: 2000 }
        markdown: { type: string, maxLength: 200000 }
        content_status: { $ref: '#/components/schemas/ContentStatus' }
        content_class: { type: string, maxLength: 32 }
        content_type_id: { type: [string, 'null'], format: uuid }
        category_id: { type: [string, 'null'], format: uuid }
        tag_ids: { type: array, maxItems: 50, items: { type: string, format: uuid } }
        keyword_ids: { type: array, maxItems: 100, items: { type: string, format: uuid } }
        language: { type: string, maxLength: 16 }
        editor_mode: { type: string, enum: [markdown, rich, split] }
        occurred_at: { type: [string, 'null'], format: date-time }
        location: { type: [string, 'null'], maxLength: 240 }
        project: { type: [string, 'null'], maxLength: 240 }
        structured_data: { type: object, maxProperties: 100 }
      minProperties: 2
    ContentStatus:
      type: string
      enum: [pending_capture, pending_parse, triage, draft, ai_queued, ai_processing, ai_review, merge_required, completed, archived, discarded]
...
    PostSourceSummary:
      type: object
      required: [id, source_type, status, captured_at]
      properties:
        id: { type: string, format: uuid }
        source_type: { type: string }
        status: { type: string }
        original_url: { type: [string, 'null'], format: uri }
        original_title: { type: [string, 'null'] }
        captured_at: { type: string, format: date-time }
...
    PostAiSummary:
      type: object
      required: [display_status, optimization_count]
      properties:
        display_status: { type: [string, 'null'] }
        optimization_count: { type: integer, minimum: 0 }
        first_optimized_at: { type: [string, 'null'], format: date-time }
        last_optimized_at: { type: [string, 'null'], format: date-time }
        latest_job_id: { type: [string, 'null'], format: uuid }
        pending_candidate_id: { type: [string, 'null'], format: uuid }
...
    TaxonomyItem:
      type: object
      required: [id, kind, name, enabled, usage_count]
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [category, tag, keyword] }
        name: { type: string, maxLength: 120 }
        description: { type: [string, 'null'], maxLength: 500 }
        parent_id: { type: [string, 'null'], format: uuid }
        aliases: { type: array, items: { type: string } }
        color: { type: [string, 'null'] }
        enabled: { type: boolean }
        stop_word: { type: boolean, default: false }
        usage_count: { type: integer, minimum: 0 }
    TaxonomyWrite:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        description: { type: [string, 'null'], maxLength: 500 }
        parent_id: { type: [string, 'null'], format: uuid }
        aliases: { type: array, maxItems: 50, uniqueItems: true, items: { type: string, maxLength: 120 } }
        color: { type: [string, 'null'], maxLength: 32 }
        enabled: { type: boolean, default: true }
        stop_word: { type: boolean, default: false }
    PostPage:
      type: object
      required: [items, next_cursor, applied_filters]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Post' } }
        next_cursor: { type: [string, 'null'] }
        applied_filters: { type: object }
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        trace_id: { type: string }